mcp-servers/saju-rag/pyproject.toml
[project]
name = "saju-rag"
version = "0.1.0"
description = "사주 명리학 지식 베이스 RAG MCP 서버"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0",
    "chromadb>=0.5",
    "openai>=1.0",
    "python-dotenv>=1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
